Got products displaying

+++ SQLInjectionTestWebsite/Shared/ProductInfo.cs
@@ -4,10 +4,10 @@ namespace SQLInjectionTestWebsite.Shared
     public class ProductInfo
     {
         public readonly string Name;
-        public readonly uint Cost;
+        public readonly float Cost;
         public readonly string Description;
 
-        public ProductInfo(string name, uint cost, string description)
+        public ProductInfo(string name, float cost, string description)
         {
             Name = name;
             Cost = cost;
